Revert "switch to queueing log embeds (so we don't get rate limited)"
This reverts commit 8a8915baea4912f5c6d32e3fc1844139c185e990.
This seems to be causing problems so we'll revert it for now, and look
into it again later.

+++ events/message_delete.go
@@ -167,9 +167,19 @@ func (bot *Bot) messageDelete(m *gateway.MessageDeleteEvent) {
 		},
 	}...)
 
-	bot.Queue(wh, "message_delete", e)
-
-	bot.DB.DeleteMessage(msg.MsgID)
+	_, err = webhook.New(wh.ID, wh.Token).ExecuteAndWait(webhook.ExecuteData{
+		AvatarURL: bot.Router.Bot.AvatarURL(),
+		Embeds:    []discord.Embed{e},
+	})
+	if err == nil {
+		bot.DB.DeleteMessage(msg.MsgID)
+	} else {
+		bot.DB.Report(db.ErrorContext{
+			Event:   "message_delete",
+			GuildID: m.GuildID,
+		}, err)
+		return
+	}
 }
 
 func hasAnyPrefixLower(s string, prefixes ...string) bool {
